requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
autoflake==2.2.1
bcrypt==4.0.1
billiard==4.1.0
black==23.9.1
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
elasticsearch==8.10.0
fastapi==0.103.1
fastapi-cache2==0.2.1
fastapi-versioning==0.10.0
flake8==6.1.0
flower==2.0.1
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
humanize==4.8.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
passlib==1.7.4
pathlib==1.0.1
pathspec==0.11.2
pendulum==2.1.2
Pillow==10.0.1
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.17.1
prometheus-fastapi-instrumentator==6.1.0
prompt-toolkit==3.0.39
pyasn1==0.5.0
pycodestyle==2.11.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyflakes==3.1.0
pyright==1.1.329
pytest==7.4.2
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
redis==4.6.0
rsa==4.9
sentry-sdk==1.31.0
six==1.16.0
sniffio==1.3.0
sqladmin==0.15.0
SQLAlchemy==2.0.21
starlette==0.27.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.7
WTForms==3.0.1
yarl==1.9.2
